moved visitors into own package
