+++ nosejob/src/main/java/com/codingrangers/nosejob/parser/ProjectParser.java
@@ -3,6 +3,7 @@ package com.codingrangers.nosejob.parser;
 import com.github.javaparser.*;
 import com.github.javaparser.ast.*;
 import com.codingrangers.nosejob.models.ProjectData;
+import com.codingrangers.nosejob.parser.visitors.ProjectVisitor;
 import com.codingrangers.nosejob.models.CodeParser;
 import org.springframework.stereotype.Component;
