+++ nosejob/src/main/java/com/codingrangers/nosejob/parser/visitors/ProjectVisitor.java
@@ -1,4 +1,4 @@
-package com.codingrangers.nosejob.parser;
+package com.codingrangers.nosejob.parser.visitors;
 
 import com.codingrangers.nosejob.models.ProjectData;
 import com.codingrangers.nosejob.models.ClassData;
